remove dummy codes from merge, seperate harmony patch
