+++ Project/1.2/Source/NewRatkin.csproj
@@ -88,6 +88,7 @@
     <Compile Include="Priest\LordJob_PrayerService.cs" />
     <Compile Include="Priest\LordToil_JoinPrayerService.cs" />
     <Compile Include="Priest\LordToil_PrayerService.cs" />
+    <Compile Include="Priest\PriestPatch.cs" />
     <Compile Include="Priest\ThinkNode_ConditionalAtDutyInteractionCell.cs" />
     <Compile Include="Priest\ThinkNode_ConditionalAtDutyRoom.cs" />
     <Compile Include="RatkinGuerrilla\IncidentWoker_RatkinThiefTunnel.cs" />
