--- Project/1.2/Source/Ratkin/CompSpawer.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using Verse;
-using Verse.AI.Group;
-using Verse.Sound;
-using System.Text;
-using System.Reflection;
-using RimWorld;
-using Harmony;
-using UnityEngine;
-using Verse.AI;
-
-namespace NewRatkin
-{
-}
