--- Project/1.2/Source/Ratkin/SPDestory.cs
@@ -1,55 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Reflection;
-using System.Linq;
-using System.Text;
-using RimWorld;
-using Verse;
-using Verse.AI;
-using Verse.Sound;
-using UnityEngine;
-using HarmonyLib;
-using RimWorld.Planet;
-
-namespace NewRatkin
-{
-    [HarmonyPatch(typeof(Pawn_ApparelTracker))]
-    [HarmonyPatch("DestroyAll")]
-    public static class DestroyGearOnDropPatch
-    {
-        [HarmonyPrefix]
-        static bool Prefix(Pawn ___pawn, ThingOwner<Apparel> ___wornApparel, DestroyMode mode)
-        {
-            DestroyApparelExtension mod = ___pawn.kindDef.GetModExtension<DestroyApparelExtension>() ?? DestroyApparelExtension.defaultValues;
-            if (mod.destroyOnlyWeapon)
-            {
-                return false;
-            }
-            return true;
-        }
-    }
-    [HarmonyPatch(typeof(Hediff_Injury))]
-    [HarmonyPatch("PostAdd")]
-    public static class Hediff_InjuryPatch
-    {
-        [HarmonyPrefix]
-        static bool Prefix(Hediff_Injury __instance, ref DamageInfo? dinfo)
-        {
-            if (__instance.comps != null)
-            {
-                for (int i = 0; i < __instance.comps.Count; i++)
-                {
-                    __instance.comps[i].CompPostPostAdd(dinfo);
-                }
-            }
-            return false;
-        }
-    }
-
-    public class DestroyApparelExtension : DefModExtension
-    {
-        public static readonly DestroyApparelExtension defaultValues = new DestroyApparelExtension();
-
-        public bool destroyOnlyWeapon = false;
-    }
-}
